fix: move utility functions to utils package
