+++ node/pkg/utils/tests/utils_test.go
@@ -1,12 +1,14 @@
-package fetcher
+package tests
 
 import (
 	"testing"
+
+	"bisonai.com/orakl/node/pkg/utils"
 )
 
 func TestAvg(t *testing.T) {
 	data := []float64{1, 2, 3, 4, 5}
-	avg := getAvg(data)
+	avg := utils.GetFloatAvg(data)
 	if avg != 3 {
 		t.Errorf("Expected 3 but got %v", avg)
 	}
@@ -14,7 +16,7 @@ func TestAvg(t *testing.T) {
 
 func TestMed(t *testing.T) {
 	data := []float64{1, 2, 3, 4, 5}
-	med := getMed(data)
+	med := utils.GetFloatMed(data)
 	if med != 3 {
 		t.Errorf("Expected 3 but got %v", med)
 	}
